refactor: performance -> show로 네이밍 변경
