+++ deploy/init.sql
@@ -1,3 +1,3 @@
 CREATE DATABASE IF NOT EXISTS user_db;
 CREATE DATABASE IF NOT EXISTS payment_db;
-CREATE DATABASE IF NOT EXISTS performance_db;
+CREATE DATABASE IF NOT EXISTS show_db;
