+++ init.sql
@@ -1,4 +1,4 @@
 CREATE DATABASE IF NOT EXISTS user_db;
 CREATE DATABASE IF NOT EXISTS payment_db;
-CREATE DATABASE IF NOT EXISTS performance_db;
+CREATE DATABASE IF NOT EXISTS show_db;
 CREATE DATABASE IF NOT EXISTS ticketing_db;
