+++ musical/src/main/java/com/truve/platform/show/ShowApplication.java
@@ -1,0 +1,12 @@
+package com.truve.platform.show;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = "com.truve.platform")
+public class ShowApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ShowApplication.class, args);
+    }
+}
