+++ musical/src/main/java/com/truve/platform/show/service/domain/constant/ShowScheduleStatus.java
@@ -1,0 +1,7 @@
+package com.truve.platform.show.service.domain.constant;
+
+public enum ShowScheduleStatus {
+	OPEN,   // 예매 가능
+	CLOSED, // 예매 마감
+	CANCELLED   // 공연 취소
+}
