+++ musical/src/main/resources/application.yml
@@ -1,0 +1,5 @@
+spring:
+  application:
+    name: show
+  profiles:
+    active: ${SPRING_PROFILES_ACTIVE:local}
