--- performance/src/main/java/com/truve/platform/performance/PerformanceApplication.java
@@ -1,12 +1,0 @@
-package com.truve.platform.performance;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication(scanBasePackages = "com.truve.platform")
-public class PerformanceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PerformanceApplication.class, args);
-    }
-}
