--- performance/src/main/java/com/truve/platform/performance/service/domain/constant/PerformanceScheduleStatus.java
@@ -1,7 +1,0 @@
-package com.truve.platform.performance.service.domain.constant;
-
-public enum PerformanceScheduleStatus {
-	OPEN,   // 예매 가능
-	CLOSED, // 예매 마감
-	CANCELLED   // 공연 취소
-}
