--- performance/src/main/java/com/truve/platform/performance/service/domain/entity/Artist.java
@@ -1,18 +1,0 @@
-package com.truve.platform.performance.service.domain.entity;
-
-import com.truve.platform.common.support.BaseEntity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "artist")
-public class Artist extends BaseEntity {
-
-	private String name;
-}
