--- performance/src/main/java/com/truve/platform/performance/service/domain/entity/PerformanceSeatGrade.java
@@ -1,44 +1,0 @@
-package com.truve.platform.performance.service.domain.entity;
-
-import com.truve.platform.common.support.BaseEntity;
-
-import jakarta.persistence.AttributeOverride;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "performance_seat_grade")
-@AttributeOverride(name = "id", column = @Column(name = "performance_seat_grade_id"))
-public class PerformanceSeatGrade extends BaseEntity {
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "performance_id", nullable = false)
-	private Performance performance;
-
-	@Column(nullable = false, length = 50)
-	private String gradeName;
-
-	@Column(nullable = false)
-	private Integer basePrice;
-
-	@Column(length = 20)
-	private String colorCode;
-
-	@Builder
-	private PerformanceSeatGrade(Performance performance, String gradeName, Integer basePrice, String colorCode) {
-		this.performance = performance;
-		this.gradeName = gradeName;
-		this.basePrice = basePrice;
-		this.colorCode = colorCode;
-	}
-}
