--- performance/src/main/resources/application-local.yml
@@ -1,29 +1,0 @@
-server:
-  port: 8083
-
-spring:
-  web:
-    encoding:
-      charset: UTF-8
-      enabled: true
-      force: true
-
-  datasource:
-    url: ${PERFORMANCE_MYSQL_URL:jdbc:mysql://localhost:3306/performance_db}
-    username: ${PERFORMANCE_MYSQL_USERNAME:root}
-    password: ${PERFORMANCE_MYSQL_PASSWORD:1234}
-    driver-class-name: com.mysql.cj.jdbc.Driver
-
-  jpa:
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        format_sql: true
-        dialect: org.hibernate.dialect.MySQLDialect
-        default_batch_fetch_size: 100
-
-logging:
-  level:
-    org.hibernate.SQL: debug
-    org.hibernate.orm.jdbc.bind: trace
