--- performance/src/main/resources/application.yml
@@ -1,5 +1,0 @@
-spring:
-  application:
-    name: performance
-  profiles:
-    active: ${SPRING_PROFILES_ACTIVE:local}
