+++ settings.gradle
@@ -3,6 +3,6 @@ include ':common'
 include ':auth-server'
 include ':api-gateway'
 include ':payment'
-include ':performance'
+include ':musical'
 include ':queue'
 include ':ticketing'
